eppic: Rename scripts to reflect validity of kernel version
Rename the eppic script to follow the format:
<eppic_script>-<valid from kernel version>_to_<valid until kernel version>.c

Kernel data structures may be modified across kernel releases and
eppic scripts hardcodes the reference to structure members. Any
change to the kernel structure will lead to failure, if the kernel
data structure changes are not propagated to the eppic scripts.

The idea to keep different versions of sample eppic scripts, which
will be valid for a range of kernel releases. This naming convention,
will help users to choose the right eppic script version for their
kernel.

Suggested-by: Atsushi Kumagai <ats-kumagai@wm.jp.nec.com>
Signed-off-by: Kamalesh Babulal <kamalesh@linux.vnet.ibm.com>